backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
langchain==0.3.26
langchain-community==0.3.27
langchain-openai==0.2.2
chromadb==0.5.11
pypdf==6.4.0
tiktoken==0.7.0
python-docx==1.1.2
python-multipart==0.0.18
pytest==8.3.3
httpx==0.27.2
